feat: link to support links from admin panel
